修改RoleVoter的邏輯,允許使用者直接對資源的功能 管理介面新增了權限設定的功能

+++ csiebug-2.0/src/csiebug/service/impl/RoleServiceImpl.java
@@ -48,13 +48,13 @@ public class RoleServiceImpl extends BasicServiceImpl implements RoleService {
 		}
 	}
 	
-	public void deleteRole(Role obj) throws ServiceException {
-		notNull(obj, "Can not delete null!");
-		
-		try {
-			roleDAO.delete(obj);
-		} catch (DAOException e) {
-			throw new ServiceException("RoleDAO delete error!", e);
+	public void deleteRoles(List<Role> obj) throws ServiceException {
+		if(obj != null && obj.size() > 0) {
+			try {
+				roleDAO.delete(obj);
+			} catch (DAOException e) {
+				throw new ServiceException("RoleDAO delete error!", e);
+			}
 		}
 	}
 }
